Framework themoviedb

+++ app/src/main/java/com/lgomez/movies/domain/model/DetailsMovie.kt
@@ -6,6 +6,6 @@ data class DetailsMovie(
     var cover: String,
     var genres: List<String>,
     var originalLanguage: String,
-    val popularity: Int,
+    val popularity: Double,
     val releaseDate: String
 )
